requirements.txt
streamlit
neo4j==5.*
python-dotenv
pymupdf
pydantic
openai
numpy
